Split file from drivers to driver.rb and schema.rb
